updates driver test and extracts settings mock
