+++ pilz_teleoperation/test/unittests/test_teleoperation_settings.py
@@ -15,22 +15,14 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import yaml
-import os
+
 import rospy
 import pytest
 from pilz_teleoperation.teleoperation_settings import TeleoperationSettings
+from param_server_mock import mocked_get_param
 from pilz_teleoperation.srv import SetTeleopSettingsRequest
 
 PKG = 'pilz_teleoperation'
-
-_package_dir = "/".join(os.path.dirname(os.path.realpath(__file__)).split("/")[:-2])
-with open(_package_dir + "/config/teleoperation_settings.yaml") as f:
-    setting_defaults = yaml.load(f.read())
-
-
-def mocked_get_param(key, *args, **kwargs):
-    return setting_defaults[key[1:]]
 
 
 velocity_changes = [("linear_velocity", .2, SetTeleopSettingsRequest.INCREASE_LINEAR_VELOCITY, .21),
